NMC-3598 - Fixing the calender and password icon not showing properly in link and mail share. Also fix NMC-4488 - iOS - Update to latest E2E encryption to keep compatibility NMC-4463 - iOS<9.6.6(3)> Sharing - Link number increases if we put the app to background or app switcher

+++ iOSClient/Networking/NCNetworking+Upload.swift
@@ -438,7 +438,9 @@ extension NCNetworking {
                                                                                "serverUrl": metadata.serverUrl,
                                                                                "account": metadata.account],
                                                                     second: 0.5)
-                    } else if error.errorCode == self.global.errorBadRequest || error.errorCode == self.global.errorUnsupportedMediaType {
+//                    } else if error.errorCode == self.global.errorBadRequest || error.errorCode == self.global.errorUnsupportedMediaType {
+                    } else if (error.errorCode == self.global.errorBadRequest || error.errorCode == self.global.errorUnsupportedMediaType) &&
+                                error.errorDescription.localizedCaseInsensitiveContains("virus") {
                         NCTransferProgress.shared.clearCountError(ocIdTransfer: metadata.ocIdTransfer)
                         self.utilityFileSystem.removeFile(atPath: self.utilityFileSystem.getDirectoryProviderStorageOcId(metadata.ocId))
                         self.database.deleteMetadataOcId(metadata.ocId)
